refactor(common): create common package for the entire application
