+++ src/main/java/woowacourse/auth/application/AuthService.java
@@ -6,7 +6,7 @@ import woowacourse.auth.domain.Token;
 import woowacourse.auth.domain.User;
 import woowacourse.auth.dto.TokenRequest;
 import woowacourse.auth.dto.TokenResponse;
-import woowacourse.exception.AuthenticationException;
+import woowacourse.common.exception.AuthenticationException;
 
 @Service
 public class AuthService {
