+++ src/main/java/woowacourse/common/config/WebConfig.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.config;
+package woowacourse.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpHeaders;
