+++ src/main/java/woowacourse/common/exception/AuthenticationException.java
@@ -1,4 +1,4 @@
-package woowacourse.exception;
+package woowacourse.common.exception;
 
 public class AuthenticationException extends RuntimeException {
