+++ src/main/java/woowacourse/common/exception/ForbiddenException.java
@@ -1,4 +1,4 @@
-package woowacourse.exception;
+package woowacourse.common.exception;
 
 public class ForbiddenException extends RuntimeException {
