+++ src/main/java/woowacourse/common/exception/InvalidCartItemException.java
@@ -1,4 +1,4 @@
-package woowacourse.exception;
+package woowacourse.common.exception;
 
 public class InvalidCartItemException extends RuntimeException {
     public InvalidCartItemException() {
