+++ src/main/java/woowacourse/common/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package woowacourse.exception;
+package woowacourse.common.exception;
 
 public class NotFoundException extends RuntimeException {
