+++ src/main/java/woowacourse/common/exception/NotInCustomerCartItemException.java
@@ -1,4 +1,4 @@
-package woowacourse.exception;
+package woowacourse.common.exception;
 
 public class NotInCustomerCartItemException extends RuntimeException {
     public NotInCustomerCartItemException() {
