+++ src/main/java/woowacourse/shoppingcart/application/CartService.java
@@ -9,8 +9,8 @@ import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.dao.ProductDao;
 import woowacourse.shoppingcart.domain.Cart;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.exception.InvalidProductException;
-import woowacourse.exception.NotInCustomerCartItemException;
+import woowacourse.common.exception.InvalidProductException;
+import woowacourse.common.exception.NotInCustomerCartItemException;
 
 @Service
 @Transactional(rollbackFor = Exception.class)
