+++ src/main/java/woowacourse/shoppingcart/application/OrderService.java
@@ -7,7 +7,7 @@ import woowacourse.shoppingcart.domain.OrderDetail;
 import woowacourse.shoppingcart.dto.request.OrderRequest;
 import woowacourse.shoppingcart.domain.Orders;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.exception.InvalidOrderException;
+import woowacourse.common.exception.InvalidOrderException;
 
 import java.util.ArrayList;
 import java.util.List;
